improved core to get more flexible direction change. Added locomotion stop check so that the controller will not take computational time when it stopped. Some clean up work.

+++ core/smartbody/smartbody-lib/src/sbm/me_ct_locomotion_speed_accelerator.hpp
@@ -59,7 +59,7 @@ public:
 	virtual ~MeCtLocomotionSpeedAccelerator();
 
 	float		update(SrVec* direction, MeCtLocomotionLimb* limb);
-	void		update_speed(float time_interval);
+	void		update_speed(double time_interval);
 	void		set_target_speed(float target_speed);
 	float		get_target_speed();
 	float		get_target_acceleration();
